src/image_ingestion_service/image_processor/requirements.txt
fastapi 
uvicorn 
boto3 
aio-pika 
pydantic 
redis
aiofiles
pillow